requirements.freeze.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-sse==2.2.0
aiomqtt==2.3.0
aiosignal==1.3.2
attrs==24.3.0
bottle==0.13.2
certifi==2024.12.14
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
easing-functions==1.0.4
fonttools==4.55.3
frozenlist==1.5.0
future==1.0.0
gevent==24.11.1
greenlet==3.1.1
idna==3.10
iso8601==2.1.0
kiwisolver==1.4.8
matplotlib==3.10.0
multidict==6.1.0
mypy==1.14.1
mypy-extensions==1.0.0
numpy==2.2.1
packaging==24.2
paho-mqtt==2.1.0
pillow==11.0.0
propcache==0.2.1
psutil==6.1.1
pygame-ce==2.5.2
pyparsing==3.2.1
pyserial==3.5
pyserial-asyncio==0.6
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
setuptools==75.6.0
six==1.17.0
sseclient==0.0.27
types-aiofiles==24.1.0.20241221
types-psutil==6.1.0.20241221
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
zope.event==5.0
zope.interface==7.2
